Transform bad URLs into ActionController::RoutingError rather than letting the URI::InvalidURIError reach the top unhandled.
If you're using subrails and doing a param-based redirect from '/foo/view/:id" to "/foo/:id", sometimes people will suck when linking to you and include errant characters such as "/foo/view/1234<".  Right now this results in a URI error right in the routes on the '<'.  Instead of this, transform it to a routing error and thus a 404.  A 404 is more useful than a server error for people that suck at linking.